:wheelchair: Add a11y tests + fixes
