+++ package.json
@@ -39,6 +39,7 @@
     "@testing-library/react": "11.2.5",
     "@testing-library/user-event": "13.0.6",
     "@types/jest": "^25.1.4",
+    "@types/jest-axe": "^3.5.1",
     "@types/node": "^12.12.38",
     "@types/react": "^16.9.27",
     "@types/react-dom": "^16.9.7",
@@ -58,6 +59,7 @@
     "eslint-plugin-react": "^7.17.0",
     "eslint-plugin-standard": "^4.0.1",
     "gh-pages": "^2.2.0",
+    "jest-axe": "^4.1.0",
     "microbundle-crl": "^0.13.10",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.4",
@@ -67,8 +69,8 @@
     "spid-smart-button": "1.1.5",
     "svgo": "^2.2.2",
     "typedoc": "^0.20.34",
-    "typescript": "3.9.9",
-    "typeface-titillium-web": "latest"
+    "typeface-titillium-web": "latest",
+    "typescript": "3.9.9"
   },
   "files": [
     "dist"
