+++ src/component.test.tsx
@@ -1,117 +1,16 @@
 import React from 'react';
-import { render, fireEvent, screen, act } from '@testing-library/react';
+import { render, fireEvent, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import { ConfigurationGET, ConfigurationPOST } from '.';
 import { providers } from './shared/providers';
 import { SPIDReactButton } from './component';
-import { Types } from './shared/types';
-
-jest.useFakeTimers();
-const TIMEOUT_VALUE = 2500;
-const defaultURL = '/myLogin/idp={{idp}}';
-
-const configurations: [ConfigurationGET, ConfigurationPOST] = [
-  { method: 'GET' },
-  { method: 'POST', fieldName: 'a' }
-];
-
-const types: Types[] = ['modal', 'dropdown'];
-
-function getAllProvidersButtons(
-  method: (ConfigurationGET | ConfigurationPOST)['method'],
-  criteria = /(Accedi a SPID con|Provider non)/
-) {
-  return screen.queryAllByRole(method === 'GET' ? 'link' : 'button', {
-    name: criteria
-  });
-}
-
-function getOnlyEnabledProviders(
-  method: (ConfigurationGET | ConfigurationPOST)['method']
-) {
-  return getAllProvidersButtons(method, /Accedi a SPID con/);
-}
-
-function getOnlyDisabledProviders(
-  method: (ConfigurationGET | ConfigurationPOST)['method']
-) {
-  return getAllProvidersButtons(method, /Provider non/);
-}
-// Sometimes it is useful to mute the console for internal warnings/error
-// just be sure to check that we aren't hiding other errors with that
-function muteConsoleWithCheck(
-  fn: Function,
-  messageToFilter: string | RegExp = ''
-) {
-  const originalError = console.error;
-  console.error = jest.fn();
-
-  const result = fn();
-
-  if (messageToFilter) {
-    expect(console.error).toHaveBeenCalledWith(
-      expect.stringMatching(messageToFilter),
-      undefined
-    );
-  } else {
-    expect(console.error).toHaveBeenCalled();
-  }
-  console.error = originalError;
-  return result;
-}
-
-type TestHelper = {
-  getMainButton: () => HTMLElement;
-  openList: () => void;
-  closeList: () => void;
-  clickProvider: (provider: HTMLElement) => boolean;
-  getEnabledProviders: () => HTMLElement[];
-  getDisabledProviders: () => HTMLElement[];
-  getProviders: () => HTMLElement[];
-};
-function createHelper(
-  type: Types,
-  method: (ConfigurationGET | ConfigurationPOST)['method']
-): TestHelper {
-  const buttonRole = type === 'dropdown' ? 'link' : 'button';
-  const getMainButton = () =>
-    screen.getByRole(buttonRole, { name: /Entra con SPID/ });
-  const openList = () => {
-    act(() => {
-      fireEvent.click(getMainButton());
-      jest.advanceTimersByTime(TIMEOUT_VALUE);
-    });
-  };
-  const closeList = () => {
-    act(() => {
-      if (type === 'modal') {
-        fireEvent.click(screen.getByRole('button', { name: /Annulla/ }));
-      } else {
-        fireEvent.click(
-          screen.getByRole(buttonRole, { name: /Entra con SPID/ })
-        );
-      }
-      jest.advanceTimersByTime(TIMEOUT_VALUE);
-    });
-  };
-  return {
-    getMainButton,
-    openList,
-    closeList,
-    clickProvider: (provider: HTMLElement) => {
-      if (method === 'POST') {
-        return muteConsoleWithCheck(() => {
-          // window.HTMLFormElement.prototype.submit = () => {};
-          return fireEvent.click(provider);
-        }, 'Error: Not implemented: HTMLFormElement.prototype.submit');
-      }
-      return fireEvent.click(provider);
-    },
-    getEnabledProviders: () => getOnlyEnabledProviders(method),
-    getDisabledProviders: () => getOnlyDisabledProviders(method),
-    getProviders: () => getAllProvidersButtons(method)
-  };
-}
+import {
+  types,
+  configurations,
+  createHelper,
+  defaultURL,
+  TestHelper,
+  muteConsoleWithCheck
+} from './test/utils';
 
 for (const type of types) {
   describe(`[${type.toUpperCase()}] Full component`, () => {
@@ -173,19 +72,6 @@ for (const type of types) {
             />
           );
 
-          helper.openList();
-
-          expect(helper.getProviders()).toHaveLength(providers.length);
-        });
-
-        it(`should render all providers`, () => {
-          render(
-            <SPIDReactButton
-              url={defaultURL}
-              configuration={configuration}
-              type={type}
-            />
-          );
           helper.openList();
 
           expect(helper.getProviders()).toHaveLength(providers.length);
@@ -285,7 +171,6 @@ for (const type of types) {
           const indexID = firstProviderAvailable.id?.trim();
           const provider = providers[indexID]!;
 
-          console.log({ indexID, provider });
           expect(onClickCallback).toHaveBeenCalledWith(
             provider,
             '/myLogin/idp=' +
